requirements.txt
Django==4.2.7
djangorestframework==3.14.0
requests==2.31.0
beautifulsoup4==4.12.2
django-cors-headers==4.3.1
django-filter==23.3